requirements.txt
Django==5.0.3
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
dj-database-url
drf-yasg==1.21.8
Faker==33.0.0
ipython==8.29.0
Markdown==3.7
pytest==8.3.5
setuptools==75.6.0
validate-docbr==1.10.0
